operator/requirements.txt
kubernetes
pick
requests
gitpython
kopf
pyyaml
jinja2
aiofiles
asyncio